+++ plugins/__init__.py
@@ -1,4 +1,5 @@
 from .analyser_plugin import PipelineAnalyser
 from .base import Plugin
+from .id_populator_plugin import IDPopulator
 from .integrity_plugin import IntegrityChecker
 from .wandb_plugin import WandbConfig, WandbPlugin
